virtapt: massive refactoring to make it work again
virtapt went through a big refactoring, for the get_uri function for
the sdk. The sdk is implemented differently now, and the single remaining
user of virtapt is "elbe check_updates".

The result was, that check update did not work anymore.

Clean it up and make <raw-key> work.

- use TmpdirFilesystem (removes os.path.join() usage)
- make constructor take an ElbeXML object
- drop the old gpg code, and add code supporting <raw-key>
- do not rely on <sources_list> and <apt_prefs>, but generate
  them using the same code that a normal elbe build uses

adjust check_update (the sole user of virtapt) to the new constructor
signature.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>