Fix ebase.text expected behavior
When passing *default* as a keyword argument to 'ebase.text', we
expect *default* to be return if the path to the node _doesn't_
exists.

There's a couple of places where the expected behavior is different.
By passing a _dict_ object like as default and passing a second
keyword argument called *key*.

Thus, here's the convention:

  1) Is *default* a _dict_ like object, i.e. does it implement the
     '__getitem__' method?
  2) If 1) is True, is there a second keyword called *key*?
  3) If 2) is True then get the item  *default[key]*.
  4) If 1) or 2) are False , then return *default*.

This is a weird implementation .. it would make much more sense to
pass *default=my_dict[key]* directly instead of *default=my_dict,
key=key*.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>