pkgutils: implement extract_pkg_changelog()
extract the changelog from an already downloaded debian package.

Special care must be taken if /usr/share/doc/<pkgname> is only
a symlink to the actual directory, where a changelog resides in another
package. In that case an Exception is thrown, and the extra package needs
to be downloaded and passed as the extra_pkg parameter.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>