Preprocess ISO images options XML tags
Check if the option passed to the src iso generation are valid with
the ISO9660 standard during the preprocessor phase.

By default when there's a violation to the standard, a warning message
is emited.  If the user provide the 'src-opts' tag with the 'strict'
attribute set to "true", then an exception is raised instead of a
warning message.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>