Better mako templates for xsdtoasciidoc
All elements and attributes should be listed in the generated
documentation for complex type.

The template used to only generate examples and descriptions for
complex types with a 'sequence' tag.  Other complex types with the
'all' and 'simpleContent' tag were left with a "no sequence" example
and description.

Now for every tag at top level, a documentation generated:

    - element
    - simpleType
    - complexType
    - group

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>