Add files protection from ChRootFilesystem's excursion.
If we modify one of the excursion's files, this modification won't be
commited since after exiting the context manager of the filesystem,
the excursion is undone.

Thus, the method ChRootFilesystem.protect takes a list of files to
protect from the excursion and returns the filesytem object.  This can
be used in conjonction with the 'with' statement, allowing
modification of the protected files.

* Example

  Here the user wants to link "/etc/resolv.conf" to a target.  But
  "/etc/resolv.conf" is usually restored by the excursion.

  Using the new 'protect' method, we can bypass the excursion.

  ``````````````````````````````````````````````````````````````````````
  my_file = "/etc/resolv.conf"
  with rfs.protect({my_file}):
     link(my_file, to_some_target)
  ``````````````````````````````````````````````````````````````````````

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>