Debianize - Form widget
A form is a smart grid.  It attaches a name to its children and
generates 2 buttons.

The confirm button when pressed will first call the sanitize_inputs
callback.  It should return a dictionary where the keys are the name
of the bad field and the value a string that describe why the field is
not good.  If all fields passed the sanitization test, then the
on_submit callback is called.

If the abort button is pressed, a QUIT signal is emitted by default
and the TUI is closed.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>