grubinstaller: enable UEFI (secure) boot with shim
This is a workaround for grub-installer being dependent on the system
where it is executed. When shim and the signed grub-efi variants are
installed, it fails to create a usable installation on the loop device.

Fix the installation by creating a debian boot environment (the name
does not matter but has to be different from BOOT) and copying the
signed shim to it.

The signed grub-efi is also available for other systems than amd64.
Enable installing them as well.

Convert the grub_fw_type parameter to simplify handling uefi/bios/hybrid,
with/without the shim fix, and to pass the grub efi install architecture.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>