log: add msgonly Formatter, and encode Formatter to use
a single formatter was used for all loggers, and added programmatically
in the logging_method decorator.

However, validation.txt and elbe-report.txt should be valid ASCIIDoc,
and should not get the context prefixed in their output.

Remove the call to setHandler from @logging_method() and add
it to all instantiations of Handlers.
Introduce msgonly_fmt and use that for validation and report.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>