pylint esoap: fix signature of ESoap.get_files()
get_files had specified 2 string parameters and erratically added
a _returns parameter to the function definition.

this mess was shadowed by "# pylint: disable=unused-argument", and
the uid was not checked using open_project()

fix it up.

No change is needed on the caller side, because the parameter was optional,
and never used.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>