esoap: return ELBE-FINISH only when the msg queue is empty
it possible, that messages do not get logged, when the project switches
away from project_busy. Since they are likely to be the actual error
messages, we really want to print them, before we exit.

So ignore the ret value, until the logging queue is empty, and only then
check whether the project is still busy, and whether ELBE-FINISH should
be generated.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>