soapclient: make wait_busy check for build success and exit if not
wait_busy would treat a build failure identical to a success. If
the build was not busy anymore, it would return success.

After the build is not busy anymore, use the newly added get_project()
method on the soap interface, to actually check the status, and return
an error code, upon failure.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>