init: switch from command_out() to do() to unhide gpg output
command_out() returns the error code, and also returns the stdout.
the return values are ignored in the respective codefragment though.

switch to do() and add allow_fail=True, because gpg seems to return error
when the key can not be validated completely.

This was added to debug a problem, where the the elbe-keyring.gpg did
not include all keys. But this somehow mitigated the problem.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>