init: set GNUPGHOME directory when generating elbe-keyring.gpg
When importing a key, gnupg will try to create a trustdb.gpg file in the
GNUPGHOME directory (~/.gnupg). When ~/.gnupg does not exist, gnupg will
refuse to work:

--------------------------------------------------------------------
18:53:50  INFO:root:gpg --no-options --no-default-keyring --keyring /initvm/image/.elbe-in/elbe-keyring --import
18:53:50  gpg: Fatal: /home/elbe/.gnupg: directory does not exist!
--------------------------------------------------------------------

The result is, that keys from the xml file are not added to the installer
keyring, and several packages can not be installed, when the key differs
from the default elbe key, which is also added via the system keyrings.

Set GNUPGHOME to a valid temporary path to make sure that gnupg has
access to GNUPGHOME and can create trustdb.gpg

While touching these lines also make sure, they are not longer than 80.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>