examples armel-rescue-busybox-cpio: use <file> -> better + workaround bug
The current buster qemu has a bug: when busybox calls
exec('/proc/self/exe'). This will try to exec qemu, which will
result in a desaster, because it expects to exec busybox.

This only happens with busybox. So this problem shows in diet mode,
finetuning, where busybox shell is used.

There is <file> now, which is the better solution. This old file should
be added to the future test cases, which can not serve as examples.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>