genlicence: switch to using ElbeProject
prepare to add pkglist as a parameter to ElbeFilesystem.write_licenses()

the pkglist is only available from an rpcaptcache, and so the genlicanse
command is modified, to start with an ElbeProject object.

This also changes the userinterface, and makes the command accept the
project directory instead of a chroot directory.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>