debianize u-boot: Add cross-build dependencies
Complete the list of u-boot build dependencies, so it can also be cross
compiled using `dpkg-buildpackage -a ${p_arch} -us -uc`.

Debian packages "which can satisfy dependencies of a package of a
different architecture" should mark themselves with:

  Multi-Arch: foreign

Correctly, this is the case for debian packages like make, gzip, flex,
bison, etc.

lzop should've marked itself similarly, but it doesn't do so.
Cross-building u-boot's debian package thus leads to the error:

  dpkg-checkbuilddeps: error: Unmet build dependencies: lzop

So force-use "lzop:native" in the u-boot build deps.  As stated in the
Multiarch HOWTO:

 "Any exceptions to the default rules are marked package:any
  or package:native in the package source."

Link: https://wiki.debian.org/Multiarch/HOWTO
Signed-off-by: Ahmed S. Darwish <a.darwish@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>