grub-legacy: Add a grubinstaller variant
Keep the setup and teardown of the loop device as in the grub2 version.
Obviously, use update-grub to update the grub configuration but tweak the
menu.lst (the grub-legacy config file) to work for most cases.

This does not set the console value from the XML that should be set in kopt
as well but would need code restructuring to get that information.

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>