debian control: Make elbe-debianize a py2 & py3 pkg
elbe-debianize might have the problem that its dependencies are satisfied
with a different python version than the rest of the elbe install.

The correct solution would be splitting up the package.  However, we want
to drop py2 support in the near future, so make this optional package
depend on both versions.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>