tests version: Introduction to Elbe unit testing
It's not completely clear, that elbepack is the correct module to be
tested, and not the installed version, for example.

Add a Test, that compares the version with the expected one.

This serves as an example on how to create new unit tests.

The steps to follow are:

1)
Create a file prefixed with 'test_' under "elbepack/tests".  The
file's suffix should reflects the name of the module to test.  In this
case, the suffix is 'version' which reflect the module
'elbepack.version'.  Thus, the final file can be found under
"elbepack/tests/test_version.py".

2)
Import the Python's unit testing framework and other standard modules
that are needed.

3)
Import the Elbe module(s) that are needed for testing.

4)
Create classes that inherit from 'unittest.TestCase', with the
prefix 'Test', followed by a short description of what is being
tested.  For example, 'TestElbepackVersion'.

5)
Create methods for each class with the prefix 'test_' for methods that
are to be tested.  Helper and wrapper methods should _not_ have the
'test_' prefix.  Test methods should reflect different cases of what
is being tested in their class.

6)
As a rule of thumb, try to only make a single assertion per test
method.  However, it's fine to put more.

7)
Tests in a class are executed in order according to their name.  Thus,
you should not expect a test to run before or after another one.

8)
Tests can share states.  States that are read-only can be put at the
class level, states that are mutable, or that need to be unique for
every test should be created in the 'setUp' method.

For example, 'expected_version' is a read-only state that can be
shared across tests.  However, 'my_list' is a mutable state that
_must_ be unique.

Don't use '__init__' to set up the tests.  Prefer 'setUp' and
'tearDown'.

9)
For skipping tests, use the '@unittest.skip*' decorators.  For tests
that are expected to fail, use the '@unittest.expectedFailure'
decorator.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>