Remove -bin dependency from -common
With Elbe 12.3, I introduced a packaging bug.
elbe-debianize is not installable right now.  It was changed to install for
py2 and py3 and depends on both python-elbe-common and python3-elbe-common.
Those depend on the python{,3}-elbe-bin packages, which conflict each other.

c1addeeb7a ("py3: debian - produce python3 in python2 packages")
introduced the *-common -> *-bin dependency.

Remove the dependency to make elbe-debianize installable again.
Let the transitively dependent packages depend on it directly.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>