fstab: Find a loop device for losetup
This fix issue #254, see <https://github.com/Linutronix/elbe/issues/254>.

According to losetup(8), detaching a loop device will not return EBUSY
if the device is busy since kernel 3.7, because of lazy device
destruction.

However, EBUSY is returned if we try to acquire a device that is busy,
even though its reference count is 0.

For this reason, letting losetup to find a proper device and returning
it would avoid such situation.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>