cdroms: Change mk_source_cdrom behavior
Instead of getting a single rfs, mk_source_cdrom() now receives a
dictionnary where the keys represented the name of a component on the
cdrom and the values are tuple of the rfs, the apt cache to used and
the list of tuple of package, version to used.

In other words, { "component-name": (rfs, cache, [(pkg, version)])}.

Also, the mirror is now directly provided by the caller.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>