elbeproject: Generate cdrom components for mk_source_cdrom
Let's reuse the package list generated by elbe_report() and pass it
to build_cdroms().  From there, the target component is generated
using the buildenv cache.

The main component is generated using the package list of debootstrap
and using the buildenv cache.

The chroot/sysroot-host components are generated using the list of packages installed
in them and using the proper apt cache.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>