dump: Change log level for archive validation
Elbe will detect when a file, coming from an archive, is truncated.
However, this leads to false positive errors when the user is
expecting to have this file truncated.  For example, the user could
run a script, part of the archive, and then delete it in the
finetuning section.

Since the validation file is quite small, it's easy to the eyes to see
if there was an error.  Hence, instead of generating errors and thus
failing the build, Elbe will emit warnings.  Consequently, it's the
user responsibility to ensure that the warnings in the validation file
are in accordance with their usage of the finetuning section.

In the future, sub-tags could be added to the 'archive' tag in order
to keep the build fail on error.  Those sub-tags would give a way for
the user to cherry-pick which files to expect to have been truncated.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>