hdimg: Use losetup wrapper for grubinstaller97
Instead of using fixed loop device /dev/loop0 and /dev/poop0,
grubinstaller uses its losetup wrapper to dynamicaly get a device.
This is because the fixed device could be busy, even though it has
been detached.

Also add an exception catching and log it.  This is because the
finally block can shadow an exception throwed in the try block.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>