fstab: retry losetup, when it with EAGAIN
After deleting a loppback device, it might still have allocated pages,
and may not be ready to be reinitialized. losetup returns 1 in
this case.

The issue seems to be fixed by this patch:
https://lore.kernel.org/patchwork/patch/1075391/

which has not found its way into the current debian kernel yet.

However, since the patch suggest this is about dirty pages, we call
sync, and retry.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>