pbuilder: Adjust create_pbuilder and pdebuild_build
Create a build environment for crossbuilding packages if the 'elbe pbuilder
--cross create' command is called. Now you can run 'elbe pbuilder --cross build'
to use the environment for crossbuilding. Change the create_pbuilder function
that it will create the correct chroot environment. Add the function
pbuilder_write_cross_config to write the correct config file for pbuilder.
Change the pbuilder_write_repo_hook function to take care of the different
directory names from the pbuilder environment.
Revise the pdebuild_build function to make sure that in case of --cross used
with the build command pbuilder runs with the '--host-arch' option. Throw an
error if the pbuilder environment was created with the --cross option but the
build command gets called without the --cross option and vice versa.

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>