init init-elbe.sh.mako: Handle initvm/finetuning tag
This allows execution of shell commands in the target.

The heredoc allows for multi-line commands.  For example:
----------------------------------------------------------------------
<command>
	cat <<EOF >> /etc/hosts
	127.0.1.1	thishost.localdomain	thishost
	EOF
</command>
----------------------------------------------------------------------

All commands are cated to the file /buildenv/tmp/elbe-initvm using an
heredoc.  The special heredoc delimiter ELBE_INITVM_FINETUNING_EOF is
used to avoid conflict with user heredoc.  Commands are then executed
in target.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>