filesystem: Add doctests
Most of the tests are trivial and ensure that that all code paths are
executed and proper exceptions are thrown.

Some test are less easier to do.  For example disk_usage(), which
would requires to assume the size of small directories on a
system.

Also, to avoid clashing between files of different tests, each file
created are prefix with the method tested.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>