repomanager: Fix Pylint
23:0:   R0205: (useless-object-inheritance)
53:0:   R0205: (useless-object-inheritance)
227:33: W0613: (unused-argument)
236:40: E0602: (undefined-variable)
235:30: W0613: (unused-argument)
241:40: E0602: (undefined-variable)
240:38: W0613: (unused-argument)
12:0:   W0611: (unused-import)

Renamed component to components for RepoBase.  Multiple components can
be passed in a list.  By default the value is 'None' and this apply to
all components in the repo.

For compatiblity with old callers, passing a string is the same as
passing a list with a single value in it.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>