commands init: Ignore files for elbe-devel.tar.bz2
TLDR; Unnecessary files are copied into the initvm from the
development directory.  This results in a kernel panic when installing
a new initvm with '--devel'.

--

When creating a new initvm with the '--devel' flag, a compressed
archive called 'elbe-devel.tar.bz2' is created from the development
directory.  That archive is cpio into the initrd of the initvm
installer and, at a late stage, is decompressed into the filesystem of
the initvm.  Resulting in a exact copy of the development directory.

However, this caused me a headache yesterday when I was unable to
generate a new initvm.

This is QEMU complaining about the initrd is taking all the RAM:
----------------------------------------------------------------------
/usr/bin/qemu-system-x86_64
qemu: initrd is too large, cannot support.(max: 1073577983, need 1081853998)
----------------------------------------------------------------------
one way to overcome this is to double the default size, up to 2Gib.

However, even if that's the case, the kernel will panic after
decompressing the initrd:
----------------------------------------------------------------------
[    0.017764] Kernel panic - not syncing: initrd too large to handle, disabling initrd (1080776424 needed, 1073469440 available)
----------------------------------------------------------------------

The culprits are the .git directory if not 'git-gc' recently.  This and
the elbe-build* that are blindly copied into the initvm.  Thus,
exclusion rules were added to the archiving of the development
directory.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>