commands: check-build: Create base for checkers
Checkers for a build should inherit from 'CheckBase' and register
themself with 'CheckBase.register(tag)' where 'tag' is the name of the
command on the command line.  As a special case, the command 'all'
will run all checkers.  Checkers should implement the 'run' method to
do their testing of the build.  The 'run' method should return 0 for
success.

Checkers can assume that their working directory is set to the build
directory before entering 'run'.

Any uncatched exception will result in a failure.  Calling
'self.fail(reason)', returning a value not equal to 0 from 'run' or
setting 'self.ret' to a value not equal to 0 will also result in a
failure.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>