commands: check-build: Add cdrom checker
This checker will validate the bin-cdrom.iso and src-cdrom*.iso of a
build.

It starts by generating a binary table using the source.xml.  This
table is then incrementally validated by inspecting the bin-cdrom.iso.
Every time that a binary is validated, sources are added to the source
table.  The source table is not component aware yet.

Then, the source table is validated the same way, by incrementally
validating sources found in the src-cdrom*.iso.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>