log: Fix async_logging()
There was a nasty bug in async_logging() that make it spin forever and
burn time on the cpu.  Basically after the pipe breaks itself, the
thread would enter an infinite loop if there's no newline terminating
the stream.

This commit fixes that by removing the epoll stuff and detecting the
broken pipe by reading the empty string from it.

Also since we're there, I augmented the atmost size up to 4096 bytes
instead of 80.  This should make logging a little more fluid.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>