initvm: create: Add COW logic for initvm
Rewrote the rules and dependencies and added support for COW images.
Also moved the generation of the initvm-tree into Python.

The base image is generated by the installer.  From this base image,
the initvm is created using COW semantic.  This allows to easily clone
the base image, but also to make a fresh initvm in an instant.

Here are the steps that Elbe does for generating the installer's initrd:

  1) gunzip the base initrd downloaded, given us a cpio archive.

  2) A temporary directory is created and a bunch of files is copied
  into it.  That directory is archived using cpio and merged with the
  previous archive.

  3) The big archive is then gzip again into the final initrd.gz.

To make things more clear, the step 2) is moved into init.py at the
end.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>