Add optional mirror for host architecture
If the primary mirror does not contain the packages for both target and host
architecture the build_host_sysroot process fails.

This patch adds support for describing in the XML configuration an additional
optional mirror containing packages for the host architecture, like:
<mirror>
<host>http://archive.ubuntu.com/ubuntu</host>
</mirror>

It also adds the architecture type in the sources.list file to avoid issues
with apt-get update.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
[torbenh: use xml and default mechanism to determine host sdk architecture
          rename hostsdk variable to hostsysroot
          make hostsyroot an attribute of BuildEnv]
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>