elbe initvm: adjust the ensure function
The 'elbe initvm ensure' command just considers if the initvm is running or not.
However, it should check if the daemon is running too. So change the ensure
command that it will check the daemon and after five minutes it will throw an
error if the daemon is still not running.

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>