pbuilder: Handle mirror's options
This is still a big mess, but I think it's more readable now.

get_apt_mirrors_and_keys() will return a tuple where the first element
is a a list of mirrors to write to /etc/apt/sources.list and the
second element is a list of GPG keys to be imported by apt.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>