xmlpreprocess: Add mirrors options preprocess rule
When <noauth> is used, emit a warning and add the option 'trusted=yes'
to all mirrors.  Some binary mirrors contain embedded options in their
URL.  This is also handled by the preprocessor that will move these
options out of the URL and create proper option nodes for the mirror.

For now, the node <noauth> is not removed from the XML because some
execution paths still use 'prj.has("noauth")'.  In the future, simply
uncomment after the TODO to get rid of noauth for good.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>