soapclient: Fix base64 encoding
In Py2, binascii.b2a_base64 returns a 'str' object, but in Py3
a 'bytes' object is returned.

The soap service apparently get confused when the data is a 'bytes'
object and will somehow add an extra 4 bytes at the end of the data.
This is the reason why cdrom upload did not work, because the upload
got corrupted.

Thus, the static method 'upload_file' defined for ClientAction can be
used by sub-classes.  This method ensures that the data is of type
'str'.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>