elbeproject: cope with incremental build without initvm node
When elbe control set_xml is used for incremental builds,
it chokes on the missing initvm node, which is expected when
there is already a buildenv.

Isolate the copy_initvmnode() code, and fall back to that, when
no initvmnode exists in the incremental build code path.

There still remains a problem with missing debootstrap packages.
But this will actually make elbe-ci test this codepath.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Olivier Dion <dion@linutronix.de>