tests: test_pylint: Parametrize pylint
Also skip some files listed in 'failure_set'.  These are the remaining
lint errors.  Hopefully, we can make this set empty and remove it.

Files that are listed in the set are simply skipped with the reason
that we're expecting them to fail.

It would be nice to be able to use 'unittest.expectedFailure', but
this function doesn't work on a 'TestCase' object.  Indeed, it marks
the test method to be expected to fail.  And since this is done before
the parametrization of the test case object, there's no possible way
to use this decorator at our advantage.  Thus, using 'self.skip' is
the closest thing that mimics what we want.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>