rpcaptcache: Rewrite mark_install_devpkgs()
The main problem with this method was that it added to the 'dev_list'
all packages that were in the 'src_list', without checking for the
source's version.

This leads to some bug for example on stretch and stretch-backports
where libext2fs-dev is from source package e2fsprogs for the latter,
but not the former.  Thus, incorrectly marking the -dev package as
installed, even though it's the source package from stretch that is
installed.

Signed-off-by: Olivier Dion <dion@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>