elbe-finetuning: Add rm_apt_source finetuning method
When LOCALMACHINE is used within an Elbe XML, a line 'deb ... http://10.0.2.2'
is appended to /etc/apt/sources.list. This could cause a warning or error if
any apt operation is done on the resulting system. Thus add rm_apt_source method
to finetuning in order to make it possible to remove specified lines from the
sources.list on the target.

Addresses to GitHub issue #235.

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>