elbe-pbuilder: Fix get_files for --pbuilder-only
If 'elbe control --pbuilder-only get_files' is called it will check if the
path starts with 'pbuilder_cross' and if pbuilder-only is True. If this
isn't the case it will skip the download. It should also examine if the
path starts with 'pbuilder'. Therefore add a check for this case in
the get_files function.

Fixes: #280

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>
Tested-by: Ralf Schlatterbeck <rsc@runtux.com>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>