debian: add "python3-" prefix to most packages
Somehow dh_install installs python modules to /usr/lib/python3
when a package name starts with "python3-".
This was not the case for elbe-daemon, elbe-bootup-check,
elbe-debianize, elbe-soap and elbe-updated. So the files were installed
to /usr/lib/python3.7/. That prevents elbe from finding the respective
command, and, for example, the daemon in the initvm can not start.

Rename all packages containing commands to python3-* and adjust depends.
Also fix the packagenames to be installed into the initvm in
elbepack/version.py and change the defaults filename in the service file.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>