validate: Amend HTTPS mirror package check for buster
apt-transport-https is not needed for HTTPS mirrors anymore with the initvm
switch to buster which happened over a year ago. This has an apt version
that has HTTPS support built-in. However, it needs ca-certificates (or a
manually installed certificate) for the certificate validation.

Amend the check to give that advise to users.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Acked-by: Torben Hohn <torben.hohn@linutronix.de>