cdroms: fix difference in pkg lists after bin-cdrom.iso rebuild
Rebuild of a bin-cdrom.iso results in different packages in the
build output. This triggers an Error in the "Full Packagelist validation".

This happens because some packages on the CDROM do not have the same
Priority set as on the debian mirror. Priority on the CDROM is taken
from the control file inside the pkg, while the priority on the debian
mirror is changed via an overrides file for some packages.

debootstrap looks at the Packages file to decide which pkgs are required.

The fix is to make sure that the generated CDROM has the exact same
Priorities set as the debian mirror:

* Extend aptpkgutils.PackageBase with priority fields.
* Fill them from both APTPAckage and XMLPackage.
* Extend the schema to allow recording the priority for a pkg.
* Actually record the new attribute into the DOM, when adding a pkg.
* add a "prio=" parameter to repomanager.RepoBase.includedeb() and
  set that everywhere.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>