test: rework collection of TestResults by implementing class ElbeTestResult
sub test result collection does not work properly because junit.TestCase
does not support more than one failure/error result in the version
shipped in debian buster. (It is supported in git, though)

Implement ElbeTestCase to accumulate the test results and allow for the
creation of separate TestCases for subTests.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>