Revert "Fix pkg list validation error on version wildcards"
This reverts commit 57c34159ed640693d477ff2e6627f871225c3648.

Wildcards are interpreted as globs, and not as regexps.

This poses problems where versions like '10.3+deb10u8'
do not match themselves, when one of the versions is
interpreted as a regexp.

>>> re.match('10.3+deb10u8', '10.3+deb10u8') is None
True

An escape would be required.

>>> re.match('10.3\+deb10u8', '10.3+deb10u8') is None
False

This problem does not happen, when the version is interpreted as
a glob. This is implemented in a subsequent commit.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>