dump: Fix pkg list validation error on version wildcards
When version is specified using wildcards (glob), the simple comparison
of versions fails, and the build fails.

Use fnmatchcase, which is the actual matcher used in pythons glob module,
to match version strings. re is not used, because + is a special character
there and versions containing + do not match themselves, when interpreted
as re.

Fixes: #276

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>