test: add special handling for ElbeTestException
ElbeTestException saves the output of the failing command.
This output generally contains newlines which are not handled
properly when they end up in an xml attribute.

The output failure_output node is declared as the stacktrace in jenkins.

Put the backtrace there, and use the stdout attribute for the output.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>