tests test_pylint: switch from command_out() to testspecific system()
Use the test specific system() implementation that stores
the command output in the Exception thrown.

Rewrite the skip code to handle Exceptions, and the
case where no exception is thrown, when an Exception
is expected.

The test should fail, if its expected to fail, but doesn't.
If the test passes, we want to remove it from the ignore list.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>