debianize: kernel: control.mako: Try lzop multiarch variant first
Since Debian 11 'Bullseye', Multi-Arch: foreign is set on lzop binary
package which leads to a build error in the pbuilder when the used
kernel package was debianized with ELBE itself.

The build dependency on pkg 'lzop' is unmet then.

Multi-Arch: foreign and ':native' seems to conflict with each other
so lets get rid of it with an OR listing in the build dependencies.

This also keeps compatibility to previous Debian versions without lzop
Multi-Arch binary pkgs.

References:
[1] https://wiki.debian.org/Multiarch/MissingRationale#Why_do_M-A:foreign_packages_not_satisfy_:native_qualified_dependencies_in_build-dependencies.3F

Signed-off-by: Steffen Kothe <steffen.kothe@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>