debianize: barebox: control.mako: Fix Multi-Arch pkgs in Build-Depends
Add ':native' free pkg variants of flex and lzop in the Build-Depends:
section to prevent unwelcome errors in the pbuilder during build due
to a possible conflict of 'Multi-Arch:foreign' and ':native' declared
binary packages.

'Multi-Arch: foreign' and ':native' seems to conflict with each other
so lets get rid of it with an OR listing in the build dependencies for
the mentioned pkgs.

This also keeps compatibility to previous Debian versions without lzop
Multi-Arch binary pkgs.

Signed-off-by: Steffen Kothe <steffen.kothe@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>
Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>