tests: simple-armhf.xml: use /vmlinux and /initrd.img links as artifacts
The actual names of the vmlinuz and initrd.img files include
version numbers, and change over time. This results in continuous
failures of the tests.

The artifact code will follow symlinks. So just use the automatically
generated links as the artifact.

The dtb file is also located in a directory containing a version number.
Copy it to /boot and then generate /dtb link mocking /vmlinuz and
/initrd.img. Make sure that the link is relative, because it's evaluated
in the context of the initvm.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>