repomanager: Add amd64 architecture to ProjectRepo
amd64 is the initvm host architecture. It is currently hardcoded
to amd64 in several other places, so we can keep this here too.

When building cross with pbuilder, some packages like qt5 build
host tools (like qmake) for amd64. These are needed in subsequent
builds, and need to be insertable into the Project Repo.

Add amd64 and also fix the string to a list of architectures.
The case when arch is 'amd64' too is covered, because the list
is converted to a set later.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>