initvm: Update to bullseye
Change the suite and url-list to install the default and big-machine
initvms from bullseye. For qemu, buster-backports shall not be used
anymore, obviously.

Also, ubuntu-keyring is not available in bullseye because Debian bug
929165 prevented ubuntu-keyring from migrating. Remove it.
People who want to build Ubuntu filesystems should install it separately
in the initvm or create it from a custom XML.

Link: https://bugs.debian.org/929165
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>